Restructure folders and add additional plays
Existing projects are moved one folder level deeper. Their relative
dependency paths are not updated and most of the projects cannot be
compiled anymore.

This is not an issue. The paths will (eventually) be updated on the
project-by-project basis, if there is a need to use an existing
project again.